Add general system for displaying pages
This is probably overengineering, but serving HTML from Azure Functions is
already overengineering so whatever.

We don't want the page endpoint to access the database when it doesn't actually
need the data, so we manually construct a TemplateContext, and hook into that to
only fetch the messages when we need to.

I also tried passing an object with a 'get' property directly to the 'Render'
method, but apparently Scriban accesses this property even when it doesn't need
the result for anything.