Merge 'list' and 'add' pages into '/' path
This just seems like better UI. Why should you need to go to a different page to
add a new item just because you're already logged in. Arguably mixing
authentication logic into the markup isn't great, but the entire system for
serving pages is intentionally kitsch, so I don't consider it a problem.